requirements.txt

fastapi>=0.68.0
uvicorn[standard]>=0.15.0
python-multipart>=0.0.5
pydantic>=1.8.0
pydantic[email]>=1.8.0
sqlalchemy>=1.4.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
easyocr>=1.6.0
opencv-python-headless>=4.5.0
PyMuPDF>=1.20.0
httpx>=0.23.0
rapidfuzz>=2.10.0
python-dotenv>=0.19.0
requests>=2.26.0